feat: app & appitem DTO

+++ src/main/java/com/javaeasybank/creditcard/service/CardAppService.java
@@ -6,7 +6,11 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.javaeasybank.common.exception.BusinessException;
+import com.javaeasybank.creditcard.dto.CardApplicationResponseDto;
+import com.javaeasybank.creditcard.dto.CreateCardApplicationRequest;
 import com.javaeasybank.creditcard.entity.CardApplication;
+import com.javaeasybank.creditcard.enums.CardApplicationStatus;
+import com.javaeasybank.creditcard.mapper.CardApplicationMapper;
 import com.javaeasybank.creditcard.repository.CardAppRepository;
 
 import lombok.RequiredArgsConstructor;
@@ -17,19 +21,33 @@ import lombok.RequiredArgsConstructor;
 public class CardAppService {
 
 	private final CardAppRepository cardAppRepository;
+	private final CardApplicationMapper cardApplicationMapper;
 	
-	public List<CardApplication> findAll() {
-		return cardAppRepository.findAll();
+	public List<CardApplicationResponseDto> findAllDto() {
+		return cardApplicationMapper.toDtoList(cardAppRepository.findAll());
 	}
 	public CardApplication findById(Integer id) {
 	    return cardAppRepository.findById(id)
 	        .orElseThrow(() -> new BusinessException("Credit card application not found."));
 	}
-	public void save(CardApplication cardApplication) {
-		cardAppRepository.save(cardApplication);
-	}
+	// 查單筆（DTO）
+		public CardApplicationResponseDto findDtoById(Integer id) {
+			return cardApplicationMapper.toDto(findById(id));
+		}
 	public void deleteById(Integer id) {
-		cardAppRepository.deleteById(id);
+		if (!cardAppRepository.existsById(id)) {
+	        throw new BusinessException("Credit card application not found.");
+	    }
+	    cardAppRepository.deleteById(id);
 	}
-	
+	//新增申請
+	public CardApplicationResponseDto create(CreateCardApplicationRequest requestDto) {
+	    CardApplication entity = cardApplicationMapper.toEntity(requestDto);
+	    //預設值
+	    entity.setStatus(CardApplicationStatus.PENDING);
+
+	    CardApplication saved = cardAppRepository.save(entity);
+
+	    return cardApplicationMapper.toDto(saved);
+	}
 }
